update token text

+++ table2txt/table2tokens.py
@@ -5,6 +5,7 @@ import numpy as np
 import random
 from multiprocessing import Pool as ProcessPool
 from table2txt.table2graph import read_tables
+import argparse
 
 MAX_NUM_TOKENS = 100
 
@@ -13,15 +14,18 @@ class CellBuffer:
         self.max_buffer_size = max_buffer_size
         self.buffer_size = 0
         self.text_buffer = []
-        self.stride_buffer = []
 
-        self.custom_attr_col_idx = '_custom_attr_col_idx'
-        self.custom_attr_text = '_custom_attr_text'
-        self.custom_attr_size = '_custom_attr_size'
+        self.custom_attr_col_idx = '_user_col_idx'
+        self.custom_attr_text = '_user_text'
+        self.custom_attr_size = '_user_size'
         
 
-    def can_add(col_idx, col_info, cell_info):
-        text = col_info['text'] + ' ' + cell_info['text'].strip() + ' , '
+    def can_add(self, col_idx, col_info, cell_info):
+        if col_info['text'] != '':
+            text = col_info['text'] + ' ( ' + cell_info['text'].strip() + ' ) . '
+        else:
+            text = cell_info['text'].strip() + ' . '
+
         cell_info[self.custom_attr_col_idx] = col_idx
         cell_info[self.custom_attr_text] = text
         tokens = text.split()
@@ -34,33 +38,43 @@ class CellBuffer:
             return False
         return True
 
-    def add(cell_info):
-        text = cell_info[self.custom_attr_text]
+    def add(self, cell_info):
         token_size = cell_info[self.custom_attr_size]
-        self.text_buffer.append(text)
         self.buffer_size += token_size
+        self.text_buffer.append(cell_info)
     
-    def can_pop():
+    def can_pop(self):
         return len(self.text_buffer) > 0
      
-    def pop(title)
+    def pop(self, title):
         assert(len(self.text_buffer) > 0)
-        out_cell_lst = [self.stride_buffer + self.text_buffer]
-        out_text = title + '   .   ' + ' '.join([a[self.custom_attr_text] for a in out_cell_lst])
-        out_meta = {'cols':[a[se;f.custom_attr_col_idx] for a in out_cell_lst]}
-        self.buffer_size = 0
-        self.stride_buffer = self.text_buffer[-1:]
-        self.text_buffer = []
+        out_text = title + '   .   ' + ' '.join([a[self.custom_attr_text] for a in self.text_buffer])
+        out_meta = {'cols':[a[self.custom_attr_col_idx] for a in self.text_buffer]}
         out_data = {
             'text':out_text,
             'meta':out_meta
         }
+        
+        self.text_buffer = self.text_buffer[-1:]
+        self.buffer_size = sum([a[self.custom_attr_size] for a in self.text_buffer]) 
+        if self.buffer_size > self.max_buffer_size:
+            self.text_buffer = []
+            self.buffer_size = 0
+         
         return out_data
+
+    def reset(self):
+        self.buffer_size = 0
+        self.stride_buffer = []
+        self.text_buffer = []
 
 class TextStrategy:
     def __init__(self):
         self.cell_buffer = CellBuffer()
         return
+    
+    def reset(self):
+        self.cell_buffer.reset()
 
     def get_text(self, row_data):
         return None
@@ -75,15 +89,15 @@ class SlidingStrategy(TextStrategy):
         for col_idx, col_info in enumerate(col_info_lst):
             cell_info = cell_lst[col_idx]
 
-            if self.cell_buffer.can_add(col_info, cell_info):
+            if self.cell_buffer.can_add(col_idx, col_info, cell_info):
                 self.cell_buffer.add(cell_info)
             else:
-                text_info = self.cell_info.pop(title)
-                text_info_lst.append(text)
+                text_info = self.cell_buffer.pop(title)
+                text_info_lst.append(text_info)
         
         if self.cell_buffer.can_pop():
-            text_info = self.cell_info.pop(title)
-            text_info_lst.append(text)
+            text_info = self.cell_buffer.pop(title)
+            text_info_lst.append(text_info)
     
         return text_info_lst
 
@@ -97,11 +111,11 @@ def init_worker(strategy_name):
     global g_strategy
     g_strategy = get_strategy(strategy_name) 
 
-def get_col_data(table)
+def get_col_data(table):
     columns = table['columns']
     col_info_lst = []
     for col_data in columns:
-        col_name = col_info['text'].strip()
+        col_name = col_data['text'].strip()
         col_name_size = col_name.split()
         col_info = {
             'text':col_name,
@@ -112,12 +126,15 @@ def get_col_data(table)
 
 def process_table(table):
     table_title = table['documentTitle'].strip()
-    col_info_lst = get_col_data()
+    col_info_lst = get_col_data(table)
     row_data = table['rows']
     table_text_lst = []
     for row_idx, row_item in enumerate(row_data):
+        g_strategy.reset()
         out_text_data = g_strategy.get_text(table_title, col_info_lst, row_item)    
-        out_text_data['meta']['row'] = row_idx
+        for text_info in out_text_data:
+            text_info['meta']['row'] = row_idx
+        
         table_text_lst.extend(out_text_data) 
     return table_text_lst
 
@@ -142,17 +159,24 @@ def get_args():
   
 def main():
     args = get_args()
-    out_table_file = os.path.join('dataset', args.dataset, args.experiment, 'token_text.jsonl')
+    exprt_dir = os.path.join('dataset', args.dataset, args.experiment)
+    if os.path.exists(exprt_dir):
+        print('[%s] already exists' % exprt_dir)
+        return
+    os.makedirs(exprt_dir)
+    out_table_file = os.path.join(exprt_dir, 'token_text.jsonl')
     f_o_tables = open(out_table_file, 'w')
-    table_lst = read_tables()
+    input_table_file = os.path.join('/home/cc/data', args.dataset, 'tables', args.table_file)
+    table_lst = read_tables(input_table_file)
     N = len(table_lst)
+    
     g_p_id = 0
     init_worker(args.strategy) 
     for table in tqdm(table_lst):
         table_id = table['tableId']
         table_text_lst = process_table(table)
         for text_info in table_text_lst:
-            write_table_info(p_id, table_id, text_info, f_o_tables)
+            write_table_info(g_p_id, table_id, text_info, f_o_tables)
             g_p_id += 1
 
     f_o_tables.close()
